% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{nSdBelowMean}
\alias{nSdBelowMean}
\title{Calculate value N std dev away from mean}
\usage{
nSdBelowMean(mn, stdv, n)
}
\arguments{
\item{mn}{Mean of distribution}

\item{stdv}{std dev of distribution}

\item{n}{number of std dev below mean to calculate}
}
\value{
the value `n` standard deviations below the mean `mn`
}
\description{
This function allows you to identify genes with loss of expression
}
\keyword{internal}
