% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{receptLoss}
\alias{receptLoss}
\title{Identify genes with expression loss}
\usage{
receptLoss(exprMatrNml, exprMatrTum, nSdBelow, minPropPerGroup)
}
\arguments{
\item{exprMatrNml}{A matrix of expression values from normal tissue. Each row
is a gene, and each column is a patient or sample. Genes should be in same
order as exprMatrTum.}

\item{exprMatrTum}{A matrix of expression values from tumor tissue. Each row
is a gene, and each column is a patient or sample. Genes should be in same
order as exprMatrNml.}

\item{nSdBelow}{The number of SD below the mean of the adjacent normal tissue
to set the boundary between tumor subgroups.}

\item{minPropPerGroup}{The number of SD below the mean of the adjacent normal
tissue to set the boundary between tumor subgroups.}
}
\description{
This function allows you to identify genes with loss of expression
}
\examples{
exprMatrNml <- matrix(abs(rnorm(100, mean = 2)), nrow=10)
exprMatrTum <- matrix(abs(rnorm(100)), nrow=10)
geneNames <- paste0(letters[1:nrow(exprMatrNml)], 1:nrow(exprMatrNml))
rownames(exprMatrNml) <- rownames(exprMatrTum) <- geneNames
nSdBelow <- 2
minPropPerGroup <- .2
rl <- receptLoss(exprMatrNml, exprMatrTum, nSdBelow, minPropPerGroup)
head(rl)
}
\keyword{expression,}
\keyword{gene}
\keyword{subgroups}
